Add basic `oq` OpenQASM domain
This is mostly a minimum viable product for an OpenQASM domain in Sphinx
(`oq`) that allows documenting OpenQASM 2 and 3 `gate` instances through
Sphinx, including cross-referencing.

Only the HTML builder is supported for now, but the output should decay
to something not entirely correct but legible in other builders.

This commit doesn't have tests because I haven't thought of a really
good way to test it yet.